feat(redis): add expiration option to cacheData
Allows setting an expiration time for cached data in Redis.
This adds an optional `exp` parameter (time in seconds) to the
`CacheArgs` type and updates the `cacheData` method to use the
`EX` option when setting the value in Redis. This enables
automatic eviction of cached data after a specified duration.